Renamed GrantsFellowships header to a common include events
