+++ apps/charterafrica/src/components/EventsGrantsFellowshipsHeader/index.js
@@ -1,0 +1,3 @@
+import EventsGrantsFellowshipsHeader from "./EventsGrantsFellowshipsHeader";
+
+export default EventsGrantsFellowshipsHeader;
