--- apps/charterafrica/src/components/GrantsFellowshipsHeader/index.js
@@ -1,3 +1,0 @@
-import GrantsFellowshipsHeader from "./GrantFellowshipHeader";
-
-export default GrantsFellowshipsHeader;
