+++ apps/charterafrica/src/pages/[...slugs].page.js
@@ -2,12 +2,12 @@ import Articles from "@/charterafrica/components/Articles";
 import CommunityPlatforms from "@/charterafrica/components/CommunityPlatforms";
 import Ecosystem from "@/charterafrica/components/Ecosystem";
 import EventsGrantsAndFellowships from "@/charterafrica/components/EventsGrantsAndFellowships";
+import EventsGrantsFellowshipsHeader from "@/charterafrica/components/EventsGrantsFellowshipsHeader";
 import Explainers from "@/charterafrica/components/Explainers";
 import FAQ from "@/charterafrica/components/FAQ";
 import FeaturedPostCard from "@/charterafrica/components/FeaturedPostCard";
 import FocalCountries from "@/charterafrica/components/FocalCountries";
 import Grantees from "@/charterafrica/components/Grantees";
-import GrantsFellowshipsHeader from "@/charterafrica/components/GrantsFellowshipsHeader";
 import GuidingPrinciples from "@/charterafrica/components/GuidingPrinciples";
 import Helpdesk from "@/charterafrica/components/Helpdesk";
 import HelpdeskPageContent from "@/charterafrica/components/HelpdeskPageContent";
@@ -68,7 +68,7 @@ function Page({ blocks }) {
       case "research":
         return <Articles {...block} key={block.slug} />;
       case "fellowships-and-grants-header":
-        return <GrantsFellowshipsHeader {...block} key={block.slug} />;
+        return <EventsGrantsFellowshipsHeader {...block} key={block.slug} />;
       case "grants":
         return <EventsGrantsAndFellowships {...block} key={block.slug} />;
       case "fellowships":
